pyproject.toml
[tool.poetry]
name = "orderstatusaiobot"
version = "0.1.0"
description = ""
authors = ["AndreyKosov <andrey.kosov.05@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.6"}
aiogram = ">=3.15.0,<3.16.0"
aiohttp = ">=3.9.0,<3.10.0"
asyncpg = ">=0.30.0,<0.31.0"
python-dotenv = ">=1.0.1,<1.1.0"
uvicorn = ">=0.32.1,<0.33.0"
sqlalchemy = ">=2.0.36,<2.1.0"
aio-pika = ">=9.5.0,<9.6.0"
greenlet = ">=3.1.1,<3.2.0"
bs4 = "^0.0.2"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
